+++ Samples~/10 - Unit Testing/Scripts/Services/IScoreService.cs
@@ -1,0 +1,9 @@
+namespace ServiceKitSamples.UnitTestingExample
+{
+	public interface IScoreService
+	{
+		int CurrentScore { get; }
+		void AddScore(int points);
+		void Reset();
+	}
+}