+++ Samples~/11 - Error Handling/Scripts/Services/ICircularA.cs
@@ -1,0 +1,7 @@
+namespace ServiceKitSamples.ErrorHandlingExample
+{
+	public interface ICircularA
+	{
+		string Name { get; }
+	}
+}