+++ Samples~/11 - Error Handling/Scripts/Services/ICircularB.cs
@@ -1,0 +1,7 @@
+namespace ServiceKitSamples.ErrorHandlingExample
+{
+	public interface ICircularB
+	{
+		string Name { get; }
+	}
+}